pyproject.toml
[tool.poetry]
name = "ipm-school"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-split-settings = "^1.2.0"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
django-cryptography = "^1.1"
django-utils-six = "^2.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.6"
pillow = "^10.0.0"
django-rest-passwordreset = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
